Make User deps pluggable
This brings it in line with current best practices, but is a rather monolithic change, and I haven't tested everything thoroughly.

Long story short: a User model may be swapped out, but because we can't know that in the beginning, there is the FQ name (which works at initialization) and a method that actually gets something we can use. These are not always interchangeable. but putting them in will help clean up the User/Profile shenanigans.