Provide workarounds for mysql unicode bugs
Note that individual tables and columns must be converted to
UTF8MB4 manually, if it hasn't been done already (it has for
the production server). If deploying in the future, it *should*
be automatic, provided you follow the DB settings template.

Fuck you, MySQL. WPI, please get a PostreSQL server.

Pretty please????