Rework conditions for displaying buttons on the event detail page
Hide a lot of the buttons if the event is closed.
Check permissions before displaying more buttons.
Show more buttons for a cancelled event than previously.
Do not show the 'Reopen' button for a cancelled event, because that button doesn't uncancel the event. That button is only for reopening a closed event. Related TODO: should make an uncancel button.