Add Billing Emails
Add functionality to send billing emails with editable subject, message, and recipients. Each billing email is associated with a specific Billing instance. The bill PDF for the event is attached to each billing email.
Also lessen the number of places in the codebase where a variable named formset is used to hold a form rather than a formset.