Initial work on Event2019
This is the initial work to support the planned launch of a new pricelist in 2019. The new pricelist dramatically changes how services related to an event must be represented in LNLDB. Instead of each category of service being split into multiple levels (such as "S2: Basic Sound Event" and "S3: Large Sound Event") of which only zero or one can be selected, the new pricelist constitutes a more a-la-carte system where multiple services of the same category can be selected. Examples of services under the new system are "Small Inputs", "Large Speakers", "Subwoofers", etc.

I am referring to the changes necessary to support the new pricelist by the name of the new event model, Event2019.

The changes in this first commit are summarized below:

- Update dependencies and update code to address breaking changes in dependencies
- Break out Event into BaseEvent and Event, using new dependency django-polymorphic
- Add Event2019
- Write massive migration to deal with all of the above
- Start changing views, forms, and templates to support Event2019 (while still supporting Event)