Add feature for members to self-report as crew for an event
From the beginning of setup for an event until 3 hours after the end of the event, members will see a button on the LNLDB landing page that will allow them to self-report as crew for the event.
When a user self-reports as crew for an event, they will show as pending in the crew hours list and their name will be prefilled in a row when editing crew hours.