Large new feature: laptop passwords control
Can now keep track of the passwords for laptops and allow users to retrieve the passwords in order to use the laptops.
Each laptop will automatically rotate its passwords (at times not dictated by LNLDB) and will transmit the new passwords to LNLDB when a rotation occurs.
A log of password retrievals and rotations is visible to users with the appropriate permission.

The purpose of this feature is to increase accountability for damage of laptops by creating a log of laptop use.