Billing changes due to WPI's transition of finances from Banner to Workday
Replace the temporary stopgap system of accepting payments that was implemented in 865f839aebea4146affba455a059dfd3b0e0901c.
Add a form that collects worktag information for an event to use for charging the client through Workday. The form requires either the user to have events.edit_org_billing permission on the organization being billed or a certain hash in the "verification" get parameter. Upon successful submission of the form, the user who submitted the form is added to the associated_users of the organization being billed.
Add a "PAY NOW" button that links to the worktag collection form to the billing emails sent to clients. The link contains the "verification" get parameter that bypasses the requirement for the user to have permissions on the organization being billed.
Update the billing email template.
Add a new events list page for events that are awaiting entry into Workday after the client has completed the worktag collection form.
Tweak events list pages.
Remove the billing_fund field from the event approval form, event review form, and event form.
Add the entered_into_workday flag to the event form.
Edit the bill templates to reflect the new payment method.
Remove old billing fund information from the org detail page.