Modify Workday changes
Add field to Event2019 to store the user who fills out the worktag collection form
Add a new events list page for events that have been marked entered into Workday but not yet marked paid. Previously these events appeared on the pending payments page, mixed with events for which the worktag collection form has not yet been filled out.
Display a message when the user is filling out the worktag collection form a second or subsequent time for the same event
Email the Treasurer when the worktag collection form is filled out a second or subsequent time for the same event