Change "Licensed" to "Passed Practical"
As stated in #515, LNL is now referring to members who complete the PIT
system as "Passed Practical" rather than "Licensed". This required
changing the "Projectionist List" page as well as the "Detailed
Projectionist" table to remove the appropriate fields.

Nothing has been removed from models, just in case this is something that
needs to be brought back in the future.