Add text to indicate a test event
This presents an initial solution to #432, by adding the text
"[TEST EVENT]" to the end of every event marked as a Test Event in the
database.

Also considered was adding the unicode "⚠" character with a tooltip.
This was not added due to the potential for misinterpreting the symbol
and for accessibility on mobile.