fix: sort `upcoming`, `incoming`, `openworkorders`, and `awaitingworkday` with earliest at top
In my mind, it makes sense to see the earliest events at the top of the `upcoming`, `incoming`, `openworkorders`, and `awaitingworkday` pages. For `upcoming` and `incoming`, the soonest events are likely the ones users will be interacting with. For `openworkorders`, and `awaitingworkday`, seeing the oldest events at the top could be a good reminder that those events are either open or awaiting workday payment, respectively.